+++ frontend/src/pages/Maps.js
@@ -1,15 +1,10 @@
 import React from 'react';
-import { GoogleMap, useLoadScript, Marker } from '../../node_modules/@react-google-maps/api';
+import { GoogleMap, useLoadScript } from '../../node_modules/@react-google-maps/api';
+import Locations from '../components/locations';
 
 const mapContainerStyle = {
 	width: '100vw',
 	height: '100vh'
-};
-
-const  libraries = ['places'];
-
-const mapId = {
-	mapId: 'ae90d8aa6c7fbaf2'
 };
 
 const center = {
@@ -19,40 +14,27 @@ const center = {
 
 const options = {
 	disableDefaultUI: true,
-	zoomControl: false,
-	getIcon: false,
-	
+	zoomControl: false
 };
 
-const locations = [	
+const noPoi = [
 	{
-		name: "Soul Kitchen",
-		location: { 
-			lat: 48.12956468120664, 
-			lng: 11.574442999814972
-		},
-	},
-	{
-		name: "Mucki & Floyd",
-		location: { 
-			lat: 48.129790,
-			lng: 11.570390
-		},
-	},
-	{
-		name: "Heyluigi",
-		location: { 
-			lat: 48.128790,
-			lng: 11.568320
-		},
-	}	  
-
+		featureType: "poi",
+		elementType: "all",
+		"stylers": [
+			{ 
+			"visibility": "off" 
+			}
+		]  
+	  }
 ];
+
+
 
 export default function Map() {
 	const { isLoaded, loadError } = useLoadScript({
 		googleMapsApiKey: process.env.REACT_APP_API_KEY,
-		libraries,
+		
 	});
 	
 	if (loadError) return "Error loading maps";
@@ -61,22 +43,18 @@ export default function Map() {
 	return(
 		<GoogleMap
 			mapContainerStyle={mapContainerStyle}
-			mapId = {'ae90d8aa6c7fbaf2'}
+			noPoi={noPoi}
 			zoom={15}
 			center={center}
 			options={options}
 			
 		>
-		{
-			locations.map(item => {
-              return (
-              <Marker key={item.name} position={item.location}/>
-              )
-            })
-         }			
+		<Locations />	
 		</GoogleMap>
 	)
 };
+
+
